dev/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlite3
kafka-python>=2.0.2
pyspark>=3.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0